Refactor and move function
